pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
target-version = ["py310"]
include = '\.pyi?$'
skip-string-normalization = true
exclude = '''
(
  /(
    | \.git
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 99
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
no_lines_before = ['STDLIB', 'LOCALFOLDER']
known_third_party = [
    'pytest',
]
known_local_folder = []
multi_line_output = 3
lines_after_imports = 2
include_trailing_comma = true
use_parentheses= true


[tool.poetry]
name = "sugaru"
version = "0.0.1"
description = "Create your own syntax stupidly simple!"
readme = "README.md"
license = "MIT"
authors = ["Dmitry Makarov <mit.makaroff@gmail.com>"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [
    { include = "sugaru" },
    { include = "sugaru/py.typed" }
]


[tool.poetry.dependencies]
python = ">=3.7"
typer = {extras = ["all"], version = "^0.9.0"}

[tool.poetry.dev-dependencies]
black = "*"
mypy = ">1.0.0"
isort = "*"
pre-commit = "*"
ruff = "*"
pytest = "*"
pytest-cov = "*"
autoflake = "*"
pytest-mock = "*"
sort-all = "*"
Faker = "*"
docformatter = "*"
PyYAML = "*"
loguru = "*"
tox = "^4.5.2"
typing-extensions = "^4.6.2"
importlib-metadata = "^6.6.0"

[tool.poetry.extras]
yaml = ["pyyaml"]

[tool.mypy]
# https://mypy.readthedocs.io/en/stable/config_file.html
python_version = "3.10"
exclude = ["examples", "sugaru/__init__.py", "sugaru/mypy.py"]
warn_unused_configs = true
ignore_missing_imports = true
disallow_incomplete_defs = true

[tool.ruff]
ignore = ["E501"]
[tool.ruff.per-file-ignores]
"__init__.py" = ["F403", "F401"]


[tool.coverage]
[tool.coverage.run]
omit = [
    '__init__.py',
]
[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = True
envlist = py37,py38,py39,py310,py311

[testenv]
deps =
    black==23.3.0
    isort==5.11.5
    ruff==0.0.263

    typer==0.9.0
    pyyaml==6.0

    pytest==7.3.1
    pytest-mock==3.7.0
    mypy > 1.0.0
    faker==18.6.0
    typing-extensions==4.6.2
    importlib-metadata==6.6.0
commands = 
    black --check sugaru/ tests/
    isort --check sugaru tests/
    ruff sugaru/ tests/
    mypy sugaru/ tests/ 
    pytest tests/ -x
setenv =
    PIP_INDEX_URL = {env:PIP_INDEX_URL:https://pypi.org/simple/}

[gh-actions]
python =
    3.7: py37
    3.8: py38
    3.9: py39
    3.10: py310
    3.11: py311
"""
